requirements.txt
# 核心依赖
pandas>=1.5.0
requests>=2.28.0

# 数据库驱动 (可选安装)
# SQLite - 内置支持，无需额外安装

# MySQL支持 (可选)
mysql-connector-python>=8.0.0

# PostgreSQL支持 (可选)
psycopg2-binary>=2.9.0

# 时间序列分析
statsmodels>=0.14.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
numpy>=1.24.0

# 其他有用的包
python-dotenv>=0.19.0  # 用于环境变量管理
